students/K3339/Filatov_Arseny/Lr1/practice_1_2/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
